backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Socket.IO
python-socketio==5.11.0

# LLM Providers
openai==1.12.0
anthropic==0.18.0
httpx==0.26.0

# ML
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
shap==0.44.1
matplotlib==3.8.2

# Image ML
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3

# Storage
minio==7.2.3
boto3==1.34.34

# Utils
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
